+++ lib/music_composition.rb
@@ -4,6 +4,7 @@ require 'music_composition/version'
 require 'music_composition/notation/pitch'
 require 'music_composition/notation/note'
 require 'music_composition/notation/duration'
+require 'music_composition/notation/measure'
 require 'music_composition/theory/interval'
 require 'music_composition/theory/scale'